src/requirements.txt
appdirs==1.4.3
Babel==1.3
blinker==1.3
click==3.3
coverage==3.7.1
decorator==3.4.0
defusedxml==0.5.0
docopt==0.6.2
dominate==2.3.1
Flask==0.10.1
Flask-Babel==0.9
Flask-Bootstrap==3.3.7.1
Flask-Login==0.3.0
Flask-Mail==0.9.0
Flask-OpenID==1.2.1
Flask-SQLAlchemy==2.0
Flask-WhooshAlchemy==0.56
Flask-WTF==0.10.2
flipflop==1.0
grip==4.3.2
guess-language==0.2
gunicorn==19.1.1
itsdangerous==0.24
Jinja2==2.7.3
Markdown==2.6.8
MarkupSafe==0.23
packaging==16.8
path-and-address==2.0.1
pbr==0.10.0
pkg-resources==0.0.0
psycopg2==2.5.4
Pygments==2.2.0
pyparsing==2.2.0
python-openid==2.2.5
python3-openid==3.1.0
pytz==2014.7
requests==2.13.0
six==1.8.0
speaklater==1.3
SQLAlchemy==0.9.7
sqlalchemy-migrate==0.9.2
SQLAlchemy-Paginator==0.1
sqlparse==0.1.11
Tempita==0.5.2
visitor==0.1.3
Werkzeug==0.9.6
Whoosh==2.6.0
WTForms==2.0.1
